Font width is now measured with floating points.

+++ trunk/src/client/java/src/ar/com/tellapic/graphics/IToolBoxManager.java
@@ -105,7 +105,7 @@ public interface IToolBoxManager {
 	 * 
 	 * @param size
 	 */
-	public void setFontPropertySize(int size);
+	public void setFontPropertySize(float size);
 	
 	/**
 	 *
